requirements.txt
paramiko~=3.1.0
pexpect~=4.8.0
setuptools~=65.5.1
rich~=13.3.3
PyYAML~=6.0
Jinja2~=3.1.2
pytest~=7.3.1
requests~=2.31.0
